feat: implement ManageApiKeyForm component and refactor ApiKeyManagementCard to use it
